third_party/blink/web_tests/flag-specific/highdpi/overflow/overflow-position-001-expected.txt

This is a testharness.js-based test.
PASS position with scrollbars
FAIL position with scrollbars + rtl assert_equals: left expected 15 but got 14.666667938232422
Harness: the test ran to completion.
